style: move config dir
